extract FindProjects
